Cleanup util and move remainder to core
