+++ index.js
@@ -1,22 +1,11 @@
 'use strict'
+var core = require('./core'),
+    merge = core.util.objectMerge
 module.exports = merge(merge({
     iterator: require('./iterator'),
     array: require('./array'),
     math: require('./math'),
     push: require('./push'),
     string: require('./string'),
-    unique: require('./unique'),
-    util: require('./util')
-  },
-  require('./core')),
-  require('./common'))
-
-function merge(result, input){
-  var prop
-  for(prop in input){
-    if(input.hasOwnProperty(prop)){
-      result[prop] = input[prop]
-    }
-  }
-  return result
-}
+    unique: require('./unique')
+  }, core), require('./common'))
