--- util/_predicateToString.js
@@ -1,10 +1,0 @@
-'use strict'
-var toString = Object.prototype.toString
-
-module.exports =
-function predicateToString(type){
-  var str = '[object '+type+']'
-  return function(value){
-    return toString.call(value) === str
-  }
-}
