--- util/index.js
@@ -1,11 +1,0 @@
-'use strict'
-
-module.exports = {
-  isFunction: require('./isFunction'),
-  isArray: require('./isArray'),
-  isString: require('./isString'),
-  isRegExp: require('./isRegExp'),
-  isNumber: require('./isNumber'),
-  isUndefined: require('./isUndefined'),
-  identity: require('./identity')
-}
