--- util/isArray.js
@@ -1,1 +1,0 @@
-module.exports = Array.isArray || require('./_predicateToString')('Array')
