--- util/isFunction.js
@@ -1,6 +1,0 @@
-'use strict'
-
-module.exports =
-function isFunction(value){
-  return typeof value === 'function'
-}
