--- util/isNumber.js
@@ -1,1 +1,0 @@
-module.exports = require('./_predicateToString')('Number')
