--- util/isUndefined.js
@@ -1,6 +1,0 @@
-'use strict'
-
-module.exports =
-function isUndefined(value){
-  return value === void 0
-}
